+++ src/components/DataFetcher/DataFetcher.js
@@ -17,10 +17,10 @@ import {
   createUrlFromArray,
   NO_FILE_NAME,
   BULK_EDIT_JOB_PROFILE_NAME,
+  requestConfiguration,
 } from '../../utils';
 
-import { DataFetcherContext } from '.';
-import { requestConfiguration } from '../../utils/multipartUpload';
+import { DataFetcherContext } from './DataFetcherContext';
 
 const {
   RUNNING,
@@ -74,8 +74,7 @@ export function getJobSplittingURL(resources, splittingURL, nonSplitting) {
   return undefined;
 }
 
-@stripesConnect
-export class DataFetcher extends Component {
+class DataFetcherComponent extends Component {
   static manifest = Object.freeze({
     jobs: {
       type: 'okapi',
@@ -251,3 +250,5 @@ export class DataFetcher extends Component {
     );
   }
 }
+
+export const DataFetcher = stripesConnect(DataFetcherComponent);